Consolidates file locations, renames login
