+++ rpg-frontend/src/App.css
@@ -4,7 +4,7 @@ body{
   background:#00b3e6;
   background: linear-gradient(#00b3e6);
   font-family: "work sans";
-  background-image: url("img/WhatsApp Image 2020-07-07 at 13.27.11.jpeg");
+  background-image: url("assets/WhatsApp Image 2020-07-07 at 13.27.11.jpeg");
   background-size: cover;
 }
